+++ Functions/SetupTeardown.ps1
@@ -130,6 +130,8 @@ function Invoke-Blocks
 
     foreach ($block in $ScriptBlock)
     {
+        if ($null -eq $block) { continue }
+
         try
         {
             . $block